python/se4-w3c/requirements.txt
selenium==4.0.0.rc1
pytest
flake8
pytest-xdist
Appium-Python-Client==2.0.0rc1